requirements.txt
google-generativeai==0.8.5
python-dotenv==1.0.1
pypdf==5.3.1
numpy==2.4.2
scikit-learn==1.8.0